Override default authenticator on HelloWordlEndpoint
In order to implement our own logic in clientId whitelisting, we set
our own authenticator, ClientIdsAuthenticator.
This authenticator rely on the GoogleEndpointAuthenticator, but contains
its own clientIds list to decide or not if we return the User entity to
the endpoint backend.
In order to let this logic be implemented, we must not define clientIds parameter
on the @Api or @ApiMethod annotation.
Otherwise the GoogleEndpointAuthenticator will not return the User entity
to the ClientIdsAuthenticator.